Remove nats_stream_forwarder
We are working under the assumption that it is not necessary in the
current cf-deployment universe where nats is used much less. If this
turns out not to be the case, we can bring it back.

Note: nats_stream_forwarder does not currently consume the nats link,
so if it is brought back, the release should be updated first.

Signed-off-by: Derek Richard <drichard@pivotal.io>