Switch log api to use internal CC endpoint
- The previous, through-the-front-door/hairpin behavior was unreliable
on GCP. This took the form of 502s from the GCP loadbalancers. We're
still unsure of the root cause, there.
- This configuration uses an internal endpoint, but does so at the
expense of https. We're switching to http because the CC's internal
https endpoint requires mutual TLS, which the log API doesn't currently
provide for this call. The Loggregator team is going to fix this, at
which point they should also remove the default, and we can remove this
line entirely.

Signed-off-by: Jesse Alford <jalford@pivotal.io>