Prevent bbs encryption key label from being overwritten in transition
- This shouldn't have previously worked. The label of the current
encryption key is stored in the database itself. This label was
previously hardcoded in cf-d and overwritten during the transition
process. We had a broken build because the encryption key the database
wanted couldn't be found. This change should prevent that. More followup
will be needed to understand how this ever worked.

[#145177333]

Signed-off-by: Jesse Alford <jalford@pivotal.io>