Re-order buildpacks
The order in this file is the same as the detection order, which matters
since some apps contain files that would match multiple buildpacks.

An example would be a rails 5.1 app. Rails51 will have both a Gemfile
and a package.json. This means both ruby and nodejs would answer 'true'
for 'bin/detect', but we know that such apps should use the ruby
buildpack.

Signed-off-by: Dave Goddard <dave@goddard.id.au>