Update consul service_ttl to keep positive DNS results longer
In some scenarios where the size of the cell is small and there are
concurrent DNS resolutions happening (e.g., running WATS in parallel),
we've noticed that the file-server.cf.internal name was not getting
resolved in time, leading to failed DNS lookups. This fixes this
scenario by making DNS cache entries live longer during the CPU spikes.

[#154809327]

Signed-off-by: Andrew Keesler <akeesler@pivotal.io>