Add bosh-lite iaas-support ops file for softlayer
In order to resolve a custom bosh-lite system domain - which is required
when deploying bosh-lite on Softlayer - this commit adds a additional
folder "softlayer" to "iaas-support". This includes the specific
ops and a README file with documentation on how to use the ops file and
why it is required.
Additionally, this commit adds interpolation tests for iaas-support ops
files (currently only the Softlayer specific ops file will be tested).

Signed-off-by: Norman Sutorius <norman.sutorius@de.ibm.com>