Increase the route registrar's cc timeout on bosh-lite
We have observed in some CI systems that the route registrar will
deregister the single instance of a cloud controller if that cloud
controller is under load and cannot respond to a healthcheck in time.

So, this commit will increase the amount of time that a registrar will
allow for the cloud controller to respond.

[#158990935]

Signed-off-by: Mike Kenyon <mkenyon@pivotal.io>