Windows 2019: move online rootfs release into its own opsfile
The Garden Windows team deploys Windows 2019 cells with two rootfs releases:
online (for internet connected environments), and offline (for internetless environments).
As of 2019, the online release name is different from the offline release name.
Therefore, we should not specify the release name in the base opsfile, but instead
require a second opsfile to choose online or offline. This will help ensure we don't
mistakenly deploy the wrong version of the rootfs release.

[#162021217]

Signed-off-by: Yael Harel <yharel@pivotal.io>