Set deployment name tag in loggregator_agent and loggr-forwarder-agent
Since the loggregator forwarder_agent is now GA in cf-deployment,
in order to set a custom deployment name for metrics, we'll have to
set the deployment name tag in the loggr-forwarder-agent properties.

We also set the loggregator_agent deployment tag property to ensure v2
envelopes have the correct deployment tag.

Co-authored-by: Edwin Xie <exie@pivotal.io>