Internal Markdown links resolve correctly on GitHub
When `deployment-guide.md` was moved below `/texts`, many of the
relative links were broken when the markdown was rendered on GitHub,
returning 404s when clicked.

This corrects the links so they render properly on GitHub. I chose to
use absolute paths (i.e. begin with a "/"), but a relative path would
have worked as well.

Two links are updated to use HTTPS instead of HTTP though this is more
for consistency than security concerns.