Deploy addons to upcoming Jammy Jellyfish stemcell
We'd like to prepare cf-deployment for the upcoming Ubuntu 22.04 Jammy
Jellyfish stemcell, which we expect to be released sometime after April
2022.

To that end, we'd like the BOSH addons (e.g. BOSH DNS) to be deployable
to the beta Jammy stemcells while we prepare the stemcell for release.

This commit should have no impact on existing deployments using Xenial
or Bionic stemcells.