removing unused q-s3.log-cache.default.cf.bosh
We cargo culted this from the ops file when we inlined it into
cf-deployment. It seems unnecessary since the default bosh dns strategy
is "smart" and clients should use the service alias.

The only consumers we expect this to affect would be someone who picked
this up in the last two weeks and started using the q-s3 domain instead
of the service alias, which we assume is a null set.