Make UAA use larger VM size
UAA needs a fair amount of CPU and memory, as a large java process. My goal here is to make all the "minimal" VMs able to run with 1-2 GB of memory and probably even a shared core, in prep for an upcoming PR to BBL to modernize the machine type.