Update capi compiled release
Use a capi compiled release built on a fresh director that will
hopefully get past some of the precompilation issues that we've been
seeing recently.

This was built in the update-stemcell-and-recompile-releases job in the
update-releases pipeline. However, there were gcs connection issues and
the job failed. However, I'd like to use the capi release uploaded by
that job anyway.

Slack threads:
* https://cloudfoundry.slack.com/archives/C02HPPYQ2/p1663606305350499
* https://cloudfoundry.slack.com/archives/C07C04W4Q/p1658530420595259