Drop unused application_ca
This is only used to signed the diego_instance_ca, yet is never
referenced as a trusted CA.

Currently this is the only instance of a CA that signs another CA, and
this complicates an already complicated cerfificate rotation process.

By removing this CA signing a CA, the certificate rotation process is
simplified.