Promote system metrics ops files
Moves the ops files for deploying system metrics agents to the `addons`
folder to indicate that they are no longer experimental. Symlinking back
to the old files in the `experimental` folder prevents this from being a
breaking change.

Also adds the new ops files to the fresh-luna and windows-cedric
fan-outs so they're tested in our pipelines.