Remove CFLINUXFS3 from default cf-deployment.yml
This contribution removes cflinuxfs3 from the default
`cf-deployments.yml`. However ops files have been added to
revert the new default behaviour.

The oprations file `use-cflinuxfs3.yml` will introduce cflinuxfs3 again.
Additionally there are the operation files `use-cflinuxfs3-with-compiled-releases.yml`
as well as `use-cflinuxfs3-with-trusted-ca-cert-for-apps.yml`.

Due to a limitation in bosh a replace statement cannot be skipped
if e.g. cflinuxfs3 fields are not present. The optional bosh
parameter (?) would create them again. Without the parameter the
templating fails. Thus the only option seems to
move parts of other ops files that interfere with cflinuxfs3 provided
field to own operation files.

This now means when using `use-trusted-ca-cert-for-apps.yml` in
combination with `use-cflinuxfs3.yml` one also must use
`use-cflinuxfs3-with-trusted-ca-cert-for-apps.yml` to bring back the old
behaviour of both ops files. When using just cflinuxfs4 using
`use-trusted-ca-cert-for-apps.yml` is sufficient on its own as it also
has no reference to cflinuxfs3 anymore. Same applies to the
behaviour of the operations file `use-compiled-releases.yml`.

Additionally this contibution adopts tests and pipelines responsible for
bumping the cflinuxfs3 stack. The version bump shall only be applied to
the operation file `use-cflinuxfs3.yml` as it now brings in the
cflinuxfs3 release which is not part of `cf-deployment.yml` anymore by
default.