CI: update infrastructure pipeline
Changes the way that bbr group acquires and releases the lock for the
env. Previously, `add_claimed` and `remove` were used, which was
creating a whole new lockfile and then deleting it.. not sure why we
wanted to do that. So this change switches to using `acquire` and
`release`, which should be a valid replacement because we only expect
there to be one lock file in this pool.