CI: Fix baba-yaga-bbr env (#1080)
Removes the network-lb-gcp plan patch from baba-yaga-bbr. This plan
patch has dubious value, and is broken with the new bbl. We should rip
it out and see if everything still works.

Changes the way that bbr group acquires and releases the lock for the
env. Previously, `add_claimed` and `remove` were used, which was
creating a whole new lockfile and then deleting it.. not sure why we
wanted to do that. So this change switches to using `acquire` and
`release`, which should be a valid replacement because we only expect
there to be one lock file in this pool.

Keeps terraform plugins when bbl-ing up the baba-yaga-bbr env. These
terraform plugins are now required (as of bbl v9) to run commands like
`bbl jumpbox-address` now, which we need to run when generating the
drats config.

See https://github.com/cloudfoundry/bosh-bootloader/issues/560 for more
info.