CI: fix cf-deployment pipeline
Comment out tasks in the lite-recreate-director job that actually try to
recreate the snitch director.

The bosh-lite-gcp plan patch is broken right now, and until it's fixed
we should not recreate the snowflake environment we've got right now.

Once https://github.com/cloudfoundry/bosh-bootloader/pull/568 is merged
and released in a bbl version we can uncomment this code again.