Populate the CAPI info fields
When the CAPI info fields are not set, the info endpoints (i.e.
`API_URL/v2/info`, `API_URL/v3/info`) return empty results akin to the
following:

```
$ cf curl /v3/info
{
   "name": "",
   "build": "",
   "version": 0,
...
}
```

We should set these properties to appropriate values. Keep in mind that
`version` expects an integer value, so we cannot set version to
"v30.9.0" and build to "" as we might like.