chore: add comments to new CAPI properties
It was called out in an ARD WG meeting that the
record-cfd-version-in-manifest task in runtime-ci may have a tough time
pattern matching on `version` and `build` keys for a find-and-replace
due to the potential prevelance of those keys in the cf-deployment
manifest.

Easiest way forward seems to be to add unique comments to the keys.
These comments can serve the dual purpose of telling maintainers not to
edit these lines manually, and to help CI pattern-match these keys.