feat: Populate version info in CAPI (#1101)
* Populate the CAPI info fields

When the CAPI info fields are not set, the info endpoints (i.e.
`API_URL/v2/info`, `API_URL/v3/info`) return empty results akin to the
following:

```
$ cf curl /v3/info
{
   "name": "",
   "build": "",
   "version": 0,
...
}
```

We should set these properties to appropriate values. Keep in mind that
`version` expects an integer value, so we cannot set version to
"v30.9.0" and build to "" as we might like.

* chore: add comments to new CAPI properties

It was called out in an ARD WG meeting that the
record-cfd-version-in-manifest task in runtime-ci may have a tough time
pattern matching on `version` and `build` keys for a find-and-replace
due to the potential prevelance of those keys in the cf-deployment
manifest.

Easiest way forward seems to be to add unique comments to the keys.
These comments can serve the dual purpose of telling maintainers not to
edit these lines manually, and to help CI pattern-match these keys.