Increase Staging RAM 1 → 2.5 GiB to avoid OOMs
The most-recent set of stemcells (Jammy 1.351+) have introduced
intermittent OOM (out-of-memory) failures when staging Cloud Foundry
applications. We believe that the error is caused by a poor interaction
between the Linux 6.5 kernel and v1 control groups (cgroups) memory
controller. Specifically, the kernel is not properly reclaiming memory.

To mitigate this, we're increasing the staging RAM, which greatly
reduces the occurrence of the OOMs. One particular app, for example,
would OOM 75% of the time during staging. After increasing the RAM
limit, staging the app succeeded every time for thirty attempts.

We doubt increasing the staging RAM limit will have a negative impact
unless the user is in the habit of restaging all their applications at
the same time. The staging cycle is short-lived, and though staging an
app will reserve a greater amount of RAM, that RAM will be released when
the staging cycle completes within minutes.

This staging RAM limit has not been updated in at least eight years. [0]

[0] https://github.com/cloudfoundry/cloud_controller_ng/blame/e8fb8f31d0687e17a71833dd685d50a3929c77e6/bosh/jobs/cloud_controller_ng/spec#L637