ci: Fix lint-cf-deployment-manifest job (#1213)
Add `aliases: true` so that this job succeeds using the latest cf-d-c-t
image, which now has ruby 3.1.x rather than ruby 2.7.x.

In ruby 2.7.x, the yaml library ran an unsafe load by default, which was
okay with aliases, but in ruby 3.1.x it runs a safe load by default. We
need to tell the safe load to explicitly allow aliases.